requirements.txt
Django==4.2.27
djangorestframework==3.15.2
django-cors-headers==4.4.0
